@@ -1,45 +1,32 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-    <NetStandardImplicitPackageVersion>1.6.1</NetStandardImplicitPackageVersion>
-    <DebugType>portable</DebugType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <GenerateFullPaths>true</GenerateFullPaths>
     <OutputType>Exe</OutputType>
-    <UserSecretsId>Rodos-Secrets</UserSecretsId>
-    <PackageTargetFallback Condition=" '$(TargetFramework)' == 'netcoreapp1.1' ">$(PackageTargetFallback);portable-net45+win8+wp8+wpa81</PackageTargetFallback>
+    <UserSecretsId>rrod-secrets</UserSecretsId>
+    <LangVersion>7.3</LangVersion>
+    <ServerGarbageCollection>true</ServerGarbageCollection>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\lib\MailKit\MailKit.csproj" />
-    <ProjectReference Include="..\..\lib\Certes\Certes.csproj" />
-    <ProjectReference Include="..\..\lib\MimeKit\MimeKit.csproj" />
-    <ProjectReference Include="..\..\lib\Portable.BouncyCastle\Portable.BouncyCastle.csproj" />
     <ProjectReference Include="..\GrainInterfaces\GrainInterfaces.csproj" />
     <ProjectReference Include="..\Grains\Grains.csproj" />
   </ItemGroup>
   
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Orleans.Core" Version="2.0.0-*" />
-    <PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="2.0.0-*" />
-    <PackageReference Include="Microsoft.Orleans.OrleansProviders" Version="2.0.0-*" />
-    <PackageReference Include="WindowsAzure.Storage" Version="7.2.1" />
-    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
+    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.DockerSecrets.Unofficial" Version="1.1.2" />
+    <PackageReference Include="Microsoft.Orleans.Server" Version="2.0.3" />
+    <PackageReference Include="Microsoft.Orleans.OrleansAzureUtils" Version="2.0.0" />
+    <PackageReference Include="Certes" Version="2.3.0" />
+    <PackageReference Include="MimeKit" Version="2.0.5" />
+    <PackageReference Include="MailKit" Version="2.0.5" />
   </ItemGroup>
-  
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <DefineConstants>$(DefineConstants);RELEASE</DefineConstants>
-  </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="OrleansHost.settings.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 </Project>
