@@ -1,63 +1,38 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-    <!--<RuntimeIdentifier>win10-x64</RuntimeIdentifier>-->
-    <NetStandardImplicitPackageVersion>1.6.1</NetStandardImplicitPackageVersion>
-    <DebugType>portable</DebugType>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <GenerateFullPaths>true</GenerateFullPaths>
     <OutputType>Exe</OutputType>
-    <UserSecretsId>Rodos-Secrets</UserSecretsId>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <UserSecretsId>rrod-secrets</UserSecretsId>
     <TypescriptCompileBlocked>true</TypescriptCompileBlocked>
-    <PackageTargetFallback Condition=" '$(TargetFramework)' == 'netcoreapp1.1' ">$(PackageTargetFallback);dnxcore50;portable-net45+win8+wp8+wpa81</PackageTargetFallback>
+    <TypeScriptToolsVersion>Latest</TypeScriptToolsVersion>
+    <LangVersion>7.3</LangVersion>
   </PropertyGroup>
   
   <ItemGroup>
-    <ProjectReference Include="..\..\lib\Certes\Certes.csproj" />
-    <ProjectReference Include="..\..\lib\Portable.BouncyCastle\Portable.BouncyCastle.csproj" />
     <ProjectReference Include="..\GrainInterfaces\GrainInterfaces.csproj" />
   </ItemGroup>
   
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.DataProtection" Version="1.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Hosting.Abstractions" Version="1.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.IISIntegration" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel.Https" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.ResponseCompression" Version="1.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.WebSockets" Version="1.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.CommandLine" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.IdentityModel.Protocols.OpenIdConnect" Version="2.1.2" />
-    <PackageReference Include="Microsoft.Orleans.Client" Version="2.0.0-*" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="1.1.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="1.1.0" />
-    <PackageReference Include="Microsoft.IdentityModel.Protocol.Extensions" Version="2.0.0-*" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="1.1.0" />
-    <PackageReference Include="IdentityServer4" Version="1.1.1" />
-    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="1.0.0" />
-    <PackageReference Include="IdentityServer4.AccessTokenValidation" Version="1.0.5" />
-    <PackageReference Include="Microsoft.AspNetCore.Http.Extensions" Version="1.1.0" />
+    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.1" />
+    <PackageReference Include="Microsoft.Orleans.Client" Version="2.0.3" />
+    <PackageReference Include="Microsoft.Orleans.OrleansAzureUtils" Version="2.0.0" />
+    <PackageReference Include="IdentityServer4" Version="2.2.0" />
+    <PackageReference Include="IdentityServer4.AspNetIdentity" Version="2.1.0" />
+    <PackageReference Include="Certes" Version="2.3.0" />
+    <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.DockerSecrets.Unofficial" Version="1.1.2" />
   </ItemGroup>
 
   <ItemGroup>
+    <None Update="ClientApp\server\ApiModel.ts">
+      <CustomToolNamespace>Models\ApiModel.cs</CustomToolNamespace>
+    </None>
     <None Update="ClientApp\server\ApiResult.ts">
       <CustomToolNamespace>Models\ApiResult.cs</CustomToolNamespace>
     </None>
-    <None Update="ClientApp\server\ClientApplication.ts">
-      <CustomToolNamespace>..\GrainInterfaces\IClientApplicationGrain.cs</CustomToolNamespace>
+    <None Update="ClientApp\server\Contact.ts">
+      <CustomToolNamespace>Models\Contact.cs</CustomToolNamespace>
     </None>
     <None Update="ClientApp\server\Counter.ts">
       <CustomToolNamespace>..\GrainInterfaces\ICounterGrain.cs</CustomToolNamespace>
