@@ -29,136 +29,136 @@ namespace Webapp.Account
             _clientStore = clientStore;
         }
 
-        public async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
-        {
-            var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            if (context?.IdP != null)
-            {
-                // this is meant to short circuit the UI and only trigger the one external IdP
-                return new LoginViewModel
-                {
-                    EnableLocalLogin = false,
-                    ReturnUrl = returnUrl,
-                    Email = context?.LoginHint,
-                    ExternalProviders = new ExternalProviderModel[] {new ExternalProviderModel { AuthenticationScheme = context.IdP } }
-                };
-            }
+        //public async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
+        //{
+        //    var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
+        //    if (context?.IdP != null)
+        //    {
+        //        // this is meant to short circuit the UI and only trigger the one external IdP
+        //        return new LoginViewModel
+        //        {
+        //            EnableLocalLogin = false,
+        //            ReturnUrl = returnUrl,
+        //            Email = context?.LoginHint,
+        //            ExternalProviders = new ExternalProvider[] {new ExternalProvider { AuthenticationScheme = context.IdP } }
+        //        };
+        //    }
 
-            var schemes = _httpContextAccessor.HttpContext.Authentication.GetAuthenticationSchemes();
+        //    var schemes = _httpContextAccessor.HttpContext.Authentication.GetAuthenticationSchemes();
 
-            var providers = schemes
-                .Where(x => x.DisplayName != null && !AccountOptions.WindowsAuthenticationSchemes.Contains(x.AuthenticationScheme))
-                .Select(x => new ExternalProviderModel
-                {
-                    DisplayName = x.DisplayName,
-                    AuthenticationScheme = x.AuthenticationScheme
-                }).ToList();
+        //    var providers = schemes
+        //        .Where(x => x.DisplayName != null && !AccountOptions.WindowsAuthenticationSchemes.Contains(x.AuthenticationScheme))
+        //        .Select(x => new ExternalProvider
+        //        {
+        //            DisplayName = x.DisplayName,
+        //            AuthenticationScheme = x.AuthenticationScheme
+        //        }).ToList();
 
-            if (AccountOptions.WindowsAuthenticationEnabled)
-            {
-                // this is needed to handle windows auth schemes
-                var windowsSchemes = schemes.Where(s => AccountOptions.WindowsAuthenticationSchemes.Contains(s.AuthenticationScheme));
-                if (windowsSchemes.Any())
-                {
-                    providers.Add(new ExternalProviderModel
-                    {
-                        AuthenticationScheme = AccountOptions.WindowsAuthenticationSchemes.First(),
-                        DisplayName = AccountOptions.WindowsAuthenticationDisplayName
-                    });
-                }
-            }
+        //    if (AccountOptions.WindowsAuthenticationEnabled)
+        //    {
+        //        // this is needed to handle windows auth schemes
+        //        var windowsSchemes = schemes.Where(s => AccountOptions.WindowsAuthenticationSchemes.Contains(s.AuthenticationScheme));
+        //        if (windowsSchemes.Any())
+        //        {
+        //            providers.Add(new ExternalProvider
+        //            {
+        //                AuthenticationScheme = AccountOptions.WindowsAuthenticationSchemes.First(),
+        //                DisplayName = AccountOptions.WindowsAuthenticationDisplayName
+        //            });
+        //        }
+        //    }
 
-            var allowLocal = true;
-            if (context?.ClientId != null)
-            {
-                var client = await _clientStore.FindEnabledClientByIdAsync(context.ClientId);
-                if (client != null)
-                {
-                    allowLocal = client.EnableLocalLogin;
+        //    var allowLocal = true;
+        //    if (context?.ClientId != null)
+        //    {
+        //        var client = await _clientStore.FindEnabledClientByIdAsync(context.ClientId);
+        //        if (client != null)
+        //        {
+        //            allowLocal = client.EnableLocalLogin;
 
-                    if (client.IdentityProviderRestrictions != null && client.IdentityProviderRestrictions.Any())
-                    {
-                        providers = providers.Where(provider => client.IdentityProviderRestrictions.Contains(provider.AuthenticationScheme)).ToList();
-                    }
-                }
-            }
+        //            if (client.IdentityProviderRestrictions != null && client.IdentityProviderRestrictions.Any())
+        //            {
+        //                providers = providers.Where(provider => client.IdentityProviderRestrictions.Contains(provider.AuthenticationScheme)).ToList();
+        //            }
+        //        }
+        //    }
 
-            return new LoginViewModel
-            {
-                AllowRememberLogin = AccountOptions.AllowRememberLogin,
-                EnableLocalLogin = allowLocal && AccountOptions.AllowLocalLogin,
-                ReturnUrl = returnUrl,
-                Email = context?.LoginHint,
-                ExternalProviders = providers.ToArray()
-            };
-        }
+        //    return new LoginViewModel
+        //    {
+        //        AllowRememberLogin = AccountOptions.AllowRememberLogin,
+        //        EnableLocalLogin = allowLocal && AccountOptions.AllowLocalLogin,
+        //        ReturnUrl = returnUrl,
+        //        Email = context?.LoginHint,
+        //        ExternalProviders = providers.ToArray()
+        //    };
+        //}
 
-        public async Task<LoginViewModel> BuildLoginViewModelAsync(LoginInputModel model)
-        {
-            var vm = await BuildLoginViewModelAsync(model.ReturnUrl);
-            vm.Email = model.Email;
-            vm.RememberLogin = model.RememberLogin;
-            return vm;
-        }
+        //public async Task<LoginViewModel> BuildLoginViewModelAsync(LoginInputModel model)
+        //{
+        //    var vm = await BuildLoginViewModelAsync(model.ReturnUrl);
+        //    vm.Email = model.Email;
+        //    vm.RememberLogin = model.RememberLogin;
+        //    return vm;
+        //}
 
-        public async Task<LogoutViewModel> BuildLogoutViewModelAsync(string logoutId)
-        {
-            var vm = new LogoutViewModel { LogoutId = logoutId, ShowLogoutPrompt = AccountOptions.ShowLogoutPrompt };
+        //public async Task<LogoutViewModel> BuildLogoutViewModelAsync(string logoutId)
+        //{
+        //    var vm = new LogoutViewModel { LogoutId = logoutId, ShowLogoutPrompt = false };
 
-            var user = await _httpContextAccessor.HttpContext.GetIdentityServerUserAsync();
-            if (user == null || user.Identity.IsAuthenticated == false)
-            {
-                // if the user is not authenticated, then just show logged out page
-                vm.ShowLogoutPrompt = false;
-                return vm;
-            }
+        //    var user = await _httpContextAccessor.HttpContext.GetIdentityServerUserAsync();
+        //    if (user == null || user.Identity.IsAuthenticated == false)
+        //    {
+        //        // if the user is not authenticated, then just show logged out page
+        //        vm.ShowLogoutPrompt = false;
+        //        return vm;
+        //    }
 
-            var context = await _interaction.GetLogoutContextAsync(logoutId);
-            if (context?.ShowSignoutPrompt == false)
-            {
-                // it's safe to automatically sign-out
-                vm.ShowLogoutPrompt = false;
-                return vm;
-            }
+        //    var context = await _interaction.GetLogoutContextAsync(logoutId);
+        //    if (context?.ShowSignoutPrompt == false)
+        //    {
+        //        // it's safe to automatically sign-out
+        //        vm.ShowLogoutPrompt = false;
+        //        return vm;
+        //    }
 
-            // show the logout prompt. this prevents attacks where the user
-            // is automatically signed out by another malicious web page.
-            return vm;
-        }
+        //    // show the logout prompt. this prevents attacks where the user
+        //    // is automatically signed out by another malicious web page.
+        //    return vm;
+        //}
 
-        public async Task<LoggedOutViewModel> BuildLoggedOutViewModelAsync(string logoutId)
-        {
-            // get context information (client name, post logout redirect URI and iframe for federated signout)
-            var logout = await _interaction.GetLogoutContextAsync(logoutId);
+        //public async Task<LoggedOutViewModel> BuildLoggedOutViewModelAsync(string logoutId)
+        //{
+        //    // get context information (client name, post logout redirect URI and iframe for federated signout)
+        //    var logout = await _interaction.GetLogoutContextAsync(logoutId);
 
-            var vm = new LoggedOutViewModel
-            {
-                AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
-                PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
-                ClientName = logout?.ClientId,
-                SignOutIframeUrl = logout?.SignOutIFrameUrl,
-                LogoutId = logoutId
-            };
+        //    var vm = new LoggedOutViewModel
+        //    {
+        //        AutomaticRedirectAfterSignOut = AccountOptions.AutomaticRedirectAfterSignOut,
+        //        PostLogoutRedirectUri = logout?.PostLogoutRedirectUri,
+        //        ClientName = logout?.ClientId,
+        //        SignOutIframeUrl = logout?.SignOutIFrameUrl,
+        //        LogoutId = logoutId
+        //    };
 
-            var user = await _httpContextAccessor.HttpContext.GetIdentityServerUserAsync();
-            if (user != null)
-            {
-                var idp = user.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
-                if (idp != null && idp != IdentityServerConstants.LocalIdentityProvider)
-                {
-                    if (vm.LogoutId == null)
-                    {
-                        // if there's no current logout context, we need to create one
-                        // this captures necessary info from the current logged in user
-                        // before we signout and redirect away to the external IdP for signout
-                        vm.LogoutId = await _interaction.CreateLogoutContextAsync();
-                    }
+        //    var user = await _httpContextAccessor.HttpContext.GetIdentityServerUserAsync();
+        //    if (user != null)
+        //    {
+        //        var idp = user.FindFirst(JwtClaimTypes.IdentityProvider)?.Value;
+        //        if (idp != null && idp != IdentityServerConstants.LocalIdentityProvider)
+        //        {
+        //            if (vm.LogoutId == null)
+        //            {
+        //                // if there's no current logout context, we need to create one
+        //                // this captures necessary info from the current logged in user
+        //                // before we signout and redirect away to the external IdP for signout
+        //                vm.LogoutId = await _interaction.CreateLogoutContextAsync();
+        //            }
 
-                    vm.ExternalAuthenticationScheme = idp;
-                }
-            }
+        //            vm.ExternalAuthenticationScheme = idp;
+        //        }
+        //    }
 
-            return vm;
-        }
+        //    return vm;
+        //}
     }
 }
