@@ -1,7 +1,9 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using Newtonsoft.Json;
+using System.ComponentModel.DataAnnotations;
 
 namespace Webapp.Account
 {
+    [JsonObject]
     public class ExternalLoginConfirmationViewModel
     {
         [Required]
