@@ -1,14 +1,14 @@
 ﻿import * as React from 'react';
 import { connect } from 'react-redux';
-import { autobind } from 'core-decorators';
 import { Grid, Row, Col, Well, Panel, PanelGroup, Button, FormGroup, Form, ControlLabel, FormControl, HelpBlock } from 'react-bootstrap';
 import { ApplicationState } from '../store';
 import * as ContactStore from '../store/Contact';
+import { ContactModel } from '../server/Contact';
+import { HeadTag } from '../lib/react-head';
 
+type ContactProps = ContactStore.ContactState & { xsrfToken: string } & typeof ContactStore.actionCreators;
 
-type ContactProps = ContactStore.ContactState & typeof ContactStore.actionCreators;
-
-const initialForm: ContactStore.ContactForm = {
+const initialForm: ContactModel = {
     firstName: '',
     lastName: '',
     email: '',
@@ -16,7 +16,7 @@ const initialForm: ContactStore.ContactForm = {
     message: ''
 }
 
-class Contact extends React.Component<ContactProps, ContactStore.ContactForm> {
+class Contact extends React.Component<ContactProps, ContactModel> {
 
     constructor(props: ContactProps) {
         super(props);
@@ -29,24 +29,26 @@ class Contact extends React.Component<ContactProps, ContactStore.ContactForm> {
             this.setState(nextProps.form)
     }
 
-    @autobind
-    handleChange(e: any) {
+    handleChange = (e: any) => {
         this.setState({ [e.target.name]: e.target.value });
     }
 
-    @autobind
-    submit(event: React.FormEvent<Form>) {
+    submit = (event: React.FormEvent<Form>) => {
         this.props.submitContactForm(this.state);
         event.preventDefault();
     }
 
     public render() {
         return <Grid>
+            <HeadTag key="title" tag="title">RROD - Contact</HeadTag>
+            <HeadTag key="meta:description" tag="meta" name="description" content="Contact page. This shows a universal webform" />
+            
             <h1>Contact us</h1>
-            <Row className="row">
+            <Row>
                 <Col md={6}>
                     <Well bsSize="sm">
-                        <Form horizontal onSubmit={this.submit}>
+                        <Form horizontal method="post" onSubmit={this.submit}>
+                            <input type="hidden" name="requestVerificationToken" defaultValue={this.props.xsrfToken} />
                             <fieldset>
                                 <legend className="text-center header">Contact</legend>
                                 <FormGroup>
@@ -95,8 +97,10 @@ class Contact extends React.Component<ContactProps, ContactStore.ContactForm> {
                     </Well>
                 </Col>
                 <Col md={6}>
-                    <Panel header={
-                        <h3>Adres</h3>}>
+                    <Panel>
+                        <Panel.Heading>
+                            <h3>Adres</h3>
+                        </Panel.Heading>
                         <div className="text-center header">
                             <div>
                                 Street address<br />
@@ -122,6 +126,8 @@ class Contact extends React.Component<ContactProps, ContactStore.ContactForm> {
 }
 
 export default connect(
-    (state: ApplicationState) => state.contact, // Selects which state properties are merged into the component's props
-    ContactStore.actionCreators                 // Selects which action creators are merged into the component's props
+    (state: ApplicationState) => {
+        return { ...state.contact, xsrfToken: state.xsrf.token }
+    }, 
+    ContactStore.actionCreators
 )(Contact);
