@@ -1,27 +1,27 @@
-import { createStore, applyMiddleware, compose, combineReducers, GenericStoreEnhancer, Store } from 'redux';
+import { createStore, applyMiddleware, compose, combineReducers,  StoreEnhancer, Store, StoreEnhancerStoreCreator, ReducersMapObject } from 'redux';
 import thunk from 'redux-thunk';
-import { routerReducer } from 'react-router-redux';
-import { ApplicationState, reducers } from './store';
-import * as AppStore from './store';
+import { routerReducer, routerMiddleware } from 'react-router-redux';
+import * as StoreModule from './store';
+import { History } from 'history';
 
-export default function configureStore(initialState?: ApplicationState) {
+export default function configureStore(history: History, initialState?: StoreModule.ApplicationState) {
     // Build middleware. These are functions that can process the actions before they reach the store.
     const windowIfDefined = typeof window === 'undefined' ? null : window as any;
     // If devTools is installed, connect to it
-    const devToolsExtension = windowIfDefined && windowIfDefined.devToolsExtension as () => GenericStoreEnhancer;
+    const devToolsExtension = windowIfDefined && windowIfDefined.__REDUX_DEVTOOLS_EXTENSION__ as () => StoreEnhancer;
     const createStoreWithMiddleware = compose(
-        applyMiddleware(thunk),
-        devToolsExtension ? devToolsExtension() : f => f
+        applyMiddleware(thunk, routerMiddleware(history)),
+        devToolsExtension ? devToolsExtension() : <S>(next: StoreEnhancerStoreCreator<S>) => next
     )(createStore);
 
     // Combine all reducers and instantiate the app-wide store instance
-    const allReducers = buildRootReducer(reducers);
-    const store = createStoreWithMiddleware(allReducers, initialState) as Store<ApplicationState>;
+    const allReducers = buildRootReducer(StoreModule.reducers);
+    const store = createStoreWithMiddleware(allReducers, initialState) as Store<StoreModule.ApplicationState>;
 
     // Enable Webpack hot module replacement for reducers
     if (module.hot) {
         module.hot.accept('./store', () => {
-            const nextRootReducer = require<typeof AppStore>('./store');
+            const nextRootReducer = require<typeof StoreModule>('./store');
             store.replaceReducer(buildRootReducer(nextRootReducer.reducers));
         });
     }
@@ -29,6 +29,6 @@ export default function configureStore(initialState?: ApplicationState) {
     return store;
 }
 
-function buildRootReducer(allReducers) {
-    return combineReducers<ApplicationState>(Object.assign({}, allReducers, { routing: routerReducer }));
+function buildRootReducer(allReducers: ReducersMapObject<StoreModule.ApplicationState>) {
+    return combineReducers<StoreModule.ApplicationState>({ ...allReducers, router: routerReducer });
 }
