@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { Router, Route, HistoryBase } from 'react-router';
-import { Layout } from './components/Layout';
+import { withRouter, Switch, Route } from 'react-router-dom'
+import Layout from './components/Layout';
 import Home from './components/Home';
 import Contact from './components/Contact';
 import Login from './components/Login';
@@ -8,18 +8,31 @@ import Logout from './components/Logout';
 import Register from './components/Register';
 import Counter from './components/Counter';
 import User from './components/User';
+import * as TransitionGroup from "react-transition-group/TransitionGroup";
+import * as CSSTransition from "react-transition-group/CSSTransition";
 
-export default <Route component={ Layout }>
-    <Route path='/' components={{ body: Home }} />
-    <Route path='/contact' components={{ body: Contact }} />
-    <Route path='/counter' components={{ body: Counter }} />
-    <Route path='/register' components={{ body: Register }} />
-    <Route path='/login' components={{ body: Login }} />
-    <Route path='/logout' components={{ body: Logout }} />
-    <Route path='/user' components={{ body: User }} />
-</Route>;
-
-// Enable Hot Module Replacement (HMR)
-if (module.hot) {
-    module.hot.accept();
-}
+export const routes = <Layout>
+    <Route 
+        render={({ location }) => { 
+            const currentKey = location.pathname.split('/')[1] || '/';
+            const timeout = { enter: 300, exit: 200 };
+            return (
+				<TransitionGroup component="div">
+                    <CSSTransition key={currentKey} timeout={timeout} classNames="fade" appear>
+                        <section className="animated-page-wrapper">
+                            <Switch location={location}>
+                                <Route exact path='/' render={() => <Home/> } />
+                                <Route path='/contact' render={() => <Contact/> } />
+                                <Route path='/counter' component={ Counter } />
+                                <Route path='/register' component={ Register } />
+                                <Route path='/login' component={ Login } />
+                                <Route path='/logout' component={ Logout } />
+                                <Route path='/user' component={ User } />
+                            </Switch>
+                        </section>
+                    </CSSTransition>
+                </TransitionGroup>);
+            }
+        }
+    />
+</Layout>;
