@@ -1,14 +1,13 @@
-// This file was generated from the Models.tst template
+﻿// This file was generated from the Models.tst template
 //
 
-export interface ApiModel { 
-    result: ApiResult;
-}
 
-export interface ApiResult { 
+
+export class ApiResult  { 
     status: string;
     success: boolean;
-    message: string;
-    errors: { [key: string]: string[]; };
-    code: number;
+    message?: string;
+    errors?: { [key: string]: string[]; };
+    code?: number;
 }
+
