@@ -1,7 +1,9 @@
-// This file was generated from the Models.tst template
+﻿// This file was generated from the Models.tst template
 //
 
-export interface UserModel { 
+
+
+export class UserModel  { 
     isAuthenticated: boolean;
     userId: string;
     email: string;
