@@ -3,17 +3,10 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Webapp.Models
 {
+    [JsonObject]
     public class SubscribeModel
     {
-        [EmailAddress]
+        [EmailAddress, Required]
         public string Email { get; set; }
     }
-
-
-    [JsonObject]
-    public class FormResponse : ApiModel
-    {
-        public string Message { get; set; }
-    }
-
 }
