@@ -9,6 +9,7 @@ using System.Linq;
 using Certes.Pkcs;
 using System.Security.Cryptography.X509Certificates;
 using Microsoft.AspNetCore.Hosting;
+using Certes.Acme.Resource;
 
 namespace Microsoft.AspNetCore.Hosting
 {
@@ -63,9 +64,9 @@ namespace Microsoft.AspNetCore.Hosting
             if (pfx != null)
             {
                 cert = new X509Certificate2(pfx, options.AcmeSettings.PfxPassword);
-                if (cert.NotAfter - DateTime.UtcNow < TimeSpan.FromDays(14))
+                if (cert.NotAfter - DateTime.UtcNow < TimeSpan.FromDays(21))
                 {
-                    // Request a new cert 14 days before the current one expires
+                    // Request a new cert 21 days before the current one expires
                     pfx = await RequestNewCertificate(domainNames, options.AcmeSettings, options.SetChallengeResponse);
                     if (pfx != null)
                     {
@@ -88,7 +89,7 @@ namespace Microsoft.AspNetCore.Hosting
 
         static async Task<byte[]> RequestNewCertificate(string[] domainNames, AcmeSettings acmeSettings, Func<string, string, Task> challengeResponseReceiver)
         {
-            using (var client = new AcmeClient(new Uri(acmeSettings.AcmeServer)))
+            using (var client = new AcmeClient(new Uri(acmeSettings.AcmeUri)))
             {
                 // Create new registration
                 var account = await client.NewRegistraton($"mailto:{acmeSettings.EmailAddress}");
@@ -130,10 +131,10 @@ namespace Microsoft.AspNetCore.Hosting
                     do
                     {
                         // Wait for ACME server to validate the identifier
-                        await Task.Delay(5000 * tryCount);
+                        await Task.Delay(5000);
                         authz = await client.GetAuthorization(httpChallenge.Location);
                     }
-                    while (authz.Data.Status == EntityStatus.Pending && ++tryCount <= 5);
+                    while (authz.Data.Status == EntityStatus.Pending && ++tryCount <= 10);
 
                     if (authz.Data.Status != EntityStatus.Valid)
                     {
