@@ -4,8 +4,10 @@
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>@ViewData["Title"] - Webapp</title>
-
+        <base href="~/" />
         <link rel="stylesheet" href="~/dist/site.css" asp-append-version="true" />
+        
+        <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
     </head>
     <body>
         @RenderBody()
